ECM ,BCM,sensors classes modified

+++ ACC.h
@@ -17,7 +17,7 @@ class ACC {
  public:
   ACC(bool i) : ignition(i) {}
   const static float velocity_of_sound = 17;  // ms
-  const static int Cruise_speed = 60;
+  const static float Cruise_speed = 60;
   const static float circumference = 209.6;  // cm
   virtual void control() = 0;
   // virtual double() = 0;
